requirements.txt
fastapi == 0.88.0
uvicorn == 0.20.0
gunicorn >= 20.1.0
starlette-prometheus >= 0.9.0
numpy >= 1.23.3
scipy >= 1.9.1
transformers == 4.25.1
pillow >= 9.2.0
timm >= 0.6.11
wget >= 3.2
zss >= 1.2.0
nltk >= 3.7
sentencepiece == 0.1.95
opencv-python == 4.6.0.66
matplotlib >= 3.6.1
dbr >= 9.4.0
pypdf2 >= 2.11.1
pymupdf == 1.18.12
opencv-contrib-python == 4.6.0.66
tf2onnx >= 1.13.0
spacy >= 3.4.3
flair >= 0.11.3
case-converter >= 1.1.0
python-doctr[torch] >= 0.6.0
truecase >= 0.0.14
jupyterlab == 3.5.1
termcolor >= 2.1.1
sentence-transformers >= 2.2.2
#tensorflow-addons == 0.19.0
#tensorflow-cpu == 2.11.0
rapidfuzz == 3.0.0
flask == 2.3.1
seaborn == 0.12.2
torch == 2.0.0
torchvision == 0.15.1
pylibdmtx == 0.1.10
opencv-python-headless == 4.6.0.66
shapely==2.0.2
